fix(tests): resolve duplicate test files and lint issues
- Remove duplicate tests/ravn/unit/ test files; tests/test_ravn/ is
  the canonical location for Ravn tests
- Merge genuinely new tests into test_cli_commands.py:
  test_single_turn_exception_calls_sys_exit, test_main_invokes_app
- Fix I001 (import sort) and F401 (unused imports) in
  test_keycloak_token_issuer.py and test_rest_pats.py
