+++ tests/test_niuu/test_keycloak_token_issuer.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 import time
-from unittest.mock import AsyncMock, MagicMock, patch
 
 import jwt
 import pytest
